Add Header to differentiate PiwigoNG from other API users
Add an header to all types of query using package_info_plus.
This add an header right before the query but the netcode being a mess it may not be applied everywhere.
Ideally this should be a dio baseoption but async code can't run in init of a class and package info plus doesn't support non async calls
A third party package with async calls seems like the worst way of implementing this but this seems like the only option